requirements.txt
fastapi
uvicorn
httpx
openai
playwright
pandas
numpy
matplotlib
pillow
pdfplumber
reportlab
scikit-learn
beautifulsoup4
lxml
python-dotenv
openpyxl
geojson
turfpy
imageio-ffmpeg
networkx
scipy
python-pptx

